Deduplicate saved-source indexing
Share save-text selection between feature and Rust indexing while retaining
their separate result and diagnostic handling. Move metrics coverage to keep
the handler module within the repository size limit.
